src/VumbaSoft.AdventureWorks.Controllers/Production/WorkOrders/WorkOrdersController.cs
using Microsoft.AspNetCore.Mvc;
using VumbaSoft.AdventureWorks.Objects;
using VumbaSoft.AdventureWorks.Services;
using VumbaSoft.AdventureWorks.Validators;
using System;

namespace VumbaSoft.AdventureWorks.Controllers.Production
{
    [Area("Production")]
    public class WorkOrdersController : ValidatedController<IWorkOrderValidator, IWorkOrderService>
    {
        public WorkOrdersController(IWorkOrderValidator validator, IWorkOrderService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create(WorkOrderView order)
        {
            if (!Validator.CanCreate(order))
                return View(order);

            Service.Create(order);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(Int32 id)
        {
            return NotEmptyView(Service.Get<WorkOrderView>(id));
        }

        [HttpGet]
        public ActionResult Edit(Int32 id)
        {
            return NotEmptyView(Service.Get<WorkOrderView>(id));
        }

        [HttpPost]
        public ActionResult Edit(WorkOrderView order)
        {
            if (!Validator.CanEdit(order))
                return View(order);

            Service.Edit(order);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(Int32 id)
        {
            return NotEmptyView(Service.Get<WorkOrderView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(Int32 id)
        {
            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
